requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.6.5
awscli==1.19.103
boto3==1.17.103
botocore==1.20.103
click==8.0.1
colorama==0.4.3
dnspython==2.1.0; python_version >= '3.6'
docutils==0.15.2
email-validator==1.1.3
faker==8.9.0
flask-cors==3.0.10
flask-login==0.5.0
flask-migrate==3.0.1
flask-sqlalchemy==2.5.1
flask-wtf==0.15.1
flask==2.0.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2; python_version >= '3.5'
itsdangerous==2.0.1
jinja2==3.0.1
jmespath==0.10.0
mako==1.1.4
markupsafe==2.0.1
pyasn1==0.4.8
pyjwt==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pyyaml==5.4.1
rsa==4.7.2
s3transfer==0.4.2
six==1.16.0
sqlalchemy==1.4.20
text-unidecode==1.3
urllib3==1.26.6
werkzeug==2.0.1
wtforms==2.3.3
